+++ src/modules/tooltip/directives/tooltip/tooltip.directive.ts
@@ -300,7 +300,11 @@ export class TooltipDirective<TData = any> implements OnChanges, OnDestroy
                                                       placement: this._options.position.placement,
                                                       offset: this._options.position.offset,
                                                       flip: true,
-                                                      mouseEvent: event,
+                                                      mouseEvent: 
+                                                      {
+                                                          ...event,
+                                                          target: this._element.nativeElement,
+                                                      },
                                                       autoUpdate: false
                                                   }))
             .then(result => applyPositionResult(result));